JCF: a save-my-work commit where I get an intermediate base class to compile
I call this a "save-my-work" commit because I've basically done the
bare minimum to abstract out much of the functionality of
HDF5DataStore, but in a frankly sloppy, non-coding-standards-compliant
kind of way (e.g., I'm using protected data members in an intermediate
base class). Still, by committing this, it'll be clearer what
refinements I make moving forward (probably const getters for the
now-only-protected member data, e.g.)
