+++ plugins/HDF5DataStore.cpp
@@ -1,7 +1,7 @@
-/**
- * This is part of the DUNE DAQ Application Framework, copyright 2020.
- * Licensing/copyright details are in the COPYING file that you should have
- * received with this code.
+/**                                                                             
+ * This is part of the DUNE DAQ Application Framework, copyright 2020.          
+ * Licensing/copyright details are in the COPYING file that you should have     
+ * received with this code.                                                     
  */
 
 #include "HDF5DataStore.hpp"
